refactor(cli): move shared command helpers to the neutral command root
