+++ turbo/apps/cli/src/commands/shared/artifact-generate.ts
@@ -1,5 +1,5 @@
 import { Command } from "commander";
-import { withErrorHandler } from "../../../lib/command/with-error-handler";
+import { withErrorHandler } from "../../lib/command/with-error-handler";
 import {
   findDesignSystem,
   findTemplate,
@@ -15,8 +15,8 @@ import {
   createHtmlArtifactAuthoringPacket,
   type HtmlArtifactKind,
 } from "./html-artifact-authoring";
-import { dispatchGenerate } from "../generate/lib/dispatch";
-import type { GenerationType } from "../generate/lib/lister";
+import { dispatchGenerate } from "../zero/generate/lib/dispatch";
+import type { GenerationType } from "../zero/generate/lib/lister";
 
 interface ArtifactOptions {
   prompt?: string;
