+++ turbo/apps/cli/src/commands/shared/image-generate.ts
@@ -1,15 +1,15 @@
 import { Command, InvalidArgumentError } from "commander";
 import chalk from "chalk";
-import { generateWebImage } from "../../../lib/api/domains/web";
-import { withErrorHandler } from "../../../lib/command/with-error-handler";
+import { generateWebImage } from "../../lib/api/domains/web";
+import { withErrorHandler } from "../../lib/command/with-error-handler";
 import { createStyledImageCompilationPacket } from "./image-style-authoring";
 import {
   findImageStyle,
   listImageStyles,
 } from "@okouai/core/resource-registry";
 import { formatRegistryListing } from "./resource-listing";
-import { dispatchGenerate } from "../generate/lib/dispatch";
-import type { GenerationType } from "../generate/lib/lister";
+import { dispatchGenerate } from "../zero/generate/lib/dispatch";
+import type { GenerationType } from "../zero/generate/lib/lister";
 
 interface ImageOptions {
   prompt?: string;
