+++ turbo/apps/cli/src/commands/shared/image-style-authoring.ts
