+++ turbo/apps/cli/src/commands/shared/sprite-authoring.ts
