+++ turbo/apps/cli/src/commands/shared/voice-generate.ts
@@ -1,9 +1,9 @@
 import { Command, Option } from "commander";
 import chalk from "chalk";
-import { generateWebVoice } from "../../../lib/api/domains/web";
-import { withErrorHandler } from "../../../lib/command/with-error-handler";
-import { dispatchGenerate } from "../generate/lib/dispatch";
-import type { GenerationType } from "../generate/lib/lister";
+import { generateWebVoice } from "../../lib/api/domains/web";
+import { withErrorHandler } from "../../lib/command/with-error-handler";
+import { dispatchGenerate } from "../zero/generate/lib/dispatch";
+import type { GenerationType } from "../zero/generate/lib/lister";
 
 interface VoiceOptions {
   prompt?: string;
