+++ turbo/apps/cli/src/commands/zero/upgrade.ts
@@ -3,7 +3,7 @@ import chalk from "chalk";
 
 import { withErrorHandler } from "../../lib/command/with-error-handler";
 import { getPlatformOrigin } from "./doctor/platform-url";
-import { planUpgradeUrl } from "./shared/billing-links";
+import { planUpgradeUrl } from "../shared/billing-links";
 
 type UpgradePlan = "pro";
