+++ turbo/apps/cli/src/commands/zero/workflow/automation/index.ts
@@ -33,7 +33,7 @@ import { getZeroChatThread } from "../../../../lib/api/domains/zero-chat";
 import { listZeroModelPolicies } from "../../../../lib/api/domains/zero-model-policies";
 import { withErrorHandler } from "../../../../lib/command/with-error-handler";
 import { decodeZeroTokenPayload } from "../../../../lib/api/zero-token";
-import { parseDurationSeconds } from "../../shared/duration";
+import { parseDurationSeconds } from "../../../shared/duration";
 import {
   resolveWorkflowRef,
   type WorkflowRefOptions,
